+++ Shared/Post/PostCell.swift
@@ -2,6 +2,7 @@ import SwiftUI
 
 struct PostCell: View {
     @EnvironmentObject var postStore: PostStore
+
     @ObservedObject var post: Post
 
     var body: some View {