+++ iOS/Settings/SettingsView.swift
@@ -1,0 +1,29 @@
+import SwiftUI
+
+struct SettingsView: View {
+    @EnvironmentObject var preferences: PreferencesModel
+    @EnvironmentObject var account: AccountModel
+
+    @Binding var isPresented: Bool
+
+    var body: some View {
+        VStack {
+            SettingsHeaderView(isPresented: $isPresented)
+            Form {
+                Section(header: Text("Login Details")) {
+                    AccountView(account: account)
+                }
+                Section(header: Text("Appearance")) {
+                    PreferencesView(preferences: preferences)
+                }
+            }
+        }
+//        .preferredColorScheme(preferences.selectedColorScheme)    // See PreferencesModel for info.
+    }
+}
+
+struct SettingsView_Previews: PreviewProvider {
+    static var previews: some View {
+        SettingsView(isPresented: .constant(true))
+    }
+}